Implement Storage class
